Refactor the modules organization
