+++ models/__init__.py
@@ -1,3 +1,3 @@
-from .clientlb import ClientLB, Request
+from .request import LoadBalancer, Request
 from .garbage import GCSTW, GCC, GCI
 from .server import ServerBaseline, ServerControl
